refactor(photos): onPhotoUpload service split
The service cannot be tested and generate error in tests

The lib part now contains the logic that we test
